+++ CHANGES.rst
@@ -45,6 +45,9 @@ optimizations.
 - Turned off GMOS-S binning restriction
 - Add GTC OSIRIS spectrograph
 - Updates for docs on adding new spectrographs.  And a bok test
+- Added a new ``pypeit_collate_1d`` tool to automatically group 1D Spectra from
+  multiple files by group and coadd them.
+- PypeIt will now add HISTORY keyword entries to FITS files.
 - `use_maskdesign` is turned off for DEIMOS LVM masks
 - a new parameter `use_user_fwhm` is added in `ExtractionPar` to allow
   the user to set their preferred fwhm
@@ -55,7 +58,6 @@ optimizations.
 - `pypeit_show_2dspec` shows traces of forced extraction and manual extraction
   with a different colors
 - Updated docs about extraction and DEIMOS
-
 
 1.3.3 (24 Feb 2021)
 -------------------